backend/requirements.txt
fastapi
python-multipart

elevenlabs
flask
python-dotenv
pydub
opencv-python 
mediapipe 
numpy 
tqdm